# Optimize server wake up