@carl-mastrangelo 2019-05-29
I'm looking for ways to make this cleaner, but I didn't come up with any.  The main issue is that stream.request() needs to be issued when the stream is created.  If it happens on the call (as is today), it's too late and the server-ELG has already gone back to sleep.  

Adding the request early means having to undo it at the application side, since it may not request it immediately (such as in a ServerCall that doesn't use our stubs).  The debouncing logic needs to know both when the real ServerCall.request() comes in, and when Listener.onMessage will be called, which is only feasibile in the ServerCallImpl  (and thus has to have a lot of invasive changes).

Ideas on how to simplify this welcome.  The latency win is pretty big in the non-direct case.